chore(ffi): move the room power levels to their own file and restructure the code
